auto-reformat old keyfile formats on load
