+++ cli/src/node/test.py
@@ -1,5 +1,4 @@
 import time
-import json
 from datetime import datetime
 
 import boto3
